requirements.txt
langchain
langchain_community
langchain_core
langchain_groq
langchain_huggingface
PyPDF2
python-dotenv
streamlit
transformers
